Moves upload panel component to dedicated folder
