+++ client/src/components/ActivityBar/ActivityBar.vue
@@ -27,7 +27,7 @@ import MultiviewPanel from "@/components/Panels/MultiviewPanel.vue";
 import NotificationsPanel from "@/components/Panels/NotificationsPanel.vue";
 import SettingsPanel from "@/components/Panels/SettingsPanel.vue";
 import ToolPanel from "@/components/Panels/ToolPanel.vue";
-import UploadPanel from "@/components/Panels/UploadPanel.vue";
+import UploadPanel from "@/components/Panels/Upload/UploadPanel.vue";
 import UserToolPanel from "@/components/Panels/UserToolPanel.vue";
 import VisualizationPanel from "@/components/Visualizations/VisualizationPanel.vue";
