+++ client/src/components/Panels/Upload/UploadPanel.vue
